+++ app/frontend/bali/components/index.js
@@ -46,6 +46,8 @@ import {
   MultiSelectController
 } from '../../../components/bali/filters/index'
 import { ColumnSelectorController } from '../../../components/bali/data_table/index'
+import { DocumentEditorController } from '../../../components/bali/document_editor/index'
+import { DocumentPageController } from '../../../components/bali/document_page/index'
 import { TreeViewItemController } from '../../../components/bali/tree_view/item/index'
 
 export { TableController } from '../../../components/bali/table/index'
@@ -92,6 +94,10 @@ export {
 
 // DataTable components
 export { ColumnSelectorController } from '../../../components/bali/data_table/index'
+
+// Document components
+export { DocumentEditorController } from '../../../components/bali/document_editor/index'
+export { DocumentPageController } from '../../../components/bali/document_page/index'
 
 // Navigation components
 export { TreeViewItemController } from '../../../components/bali/tree_view/item/index'
@@ -150,6 +156,10 @@ export function registerAll (application) {
 
   // DataTable
   application.register('column-selector', ColumnSelectorController)
+
+  // Document
+  application.register('document-editor', DocumentEditorController)
+  application.register('document-page', DocumentPageController)
 
   // Navigation
   application.register('tree-view-item', TreeViewItemController)